==> 4_Data Analysis.Rmd <==
---
title: "Analysis"
output: html_document
date: "2025-3-15"
---

```{r, include = F}
## Load packages
library(tidyverse)
library(infer)
library(oddsratio)
library(scales)
library(knitr)
library(kableExtra)
library(here)
library(yhat)
library(MOTE)
library(patchwork)
library(rlang)
library(broom)
`%+%` <- paste0

## Load data
df <- readRDS("H:\\My Drive\\Research\\Projects\\Autism Identification\\Data\\Analysis-Ready Data.rds")

## Set seed for reproducibility
set.seed(857287)

## Set functions
round <- function(x, digits = 0) sprintf("%." %+% digits %+% "f", x)

get_d <- function(x) {
  
  x1 <- df[[x]][df$self_id == T]
  x2 <- df[[x]][df$self_id == F]
  
  m1 <- mean(x1, na.rm = T)
  m2 <- mean(x2, na.rm = T)
  
  sd1 <- sd(x1, na.rm = T)
  sd2 <- sd(x2, na.rm = T)
  
  n1 <- sum(!is.na(x1))
  n2 <- sum(!is.na(x2))
  
  d <- d.ind.t(m1 = m1, m2 = m2, sd1 = sd1, sd2 = sd2, n1 = n1, n2 = n2)
  
  return(d$estimate)
  
}

n_percent <- function(n) n %+% " (" %+% percent(n / sum(n), .1) %+% ")"

mean_sd <- function(x) round(mean(x, na.rm = T), 1) %+% " (" %+% round(sd(x, na.rm = T), 1) %+% ")"

get_descriptive_count <- function(x) {
  
  full_sample <- df %>%
    count(!!sym(x)) %>%
    drop_na() %>%
    mutate(full_sample = n_percent(n)) %>%
    select(-n)
  
  by_group <- df %>%
    group_by(self_id_char) %>%
    count(!!sym(x)) %>%
    drop_na() %>%
    mutate(by_group = n_percent(n)) %>%
    pivot_wider(names_from = "self_id_char", values_from = "by_group", id_cols = all_of(x)) %>%
    select(!!sym(x), `Self-Identifying`, Diagnosed)
  
  combined <- full_join(
    full_sample, 
    by_group, 
    by = x,
    relationship = "one-to-one"
  )
  
  combined %>%
    kable() %>%
    kable_styling() %>%
    return()

}

get_descriptive_continuous <- function(x) {
  
  full_sample <- df %>%
    summarize(full_sample = mean_sd(!!sym(x))) %>%
    mutate(var = x) %>%
    select(var, full_sample)
  
  by_group <- df %>%
    group_by(self_id_char) %>%
    summarize(by_group = mean_sd(!!sym(x))) %>%
    pivot_wider(names_from = "self_id_char", values_from = "by_group") %>%
    mutate(var = x) %>%
    select(var, `Self-Identifying`, Diagnosed)
  
  combined <- full_join(
    full_sample, 
    by_group, 
    by = "var",
    relationship = "one-to-one"
  )
  
  combined %>%
    kable() %>%
    kable_styling() %>%
    return()
  
}
```

# Appendix A: Detailed Sample Description

```{r}
# Diagnosis
get_descriptive_count("self_id")

# Age
get_descriptive_continuous("age")

# Gender
get_descriptive_count("gender")
get_descriptive_count("gender_3")
get_descriptive_count("gender_4")
get_descriptive_count("gender_5")
get_descriptive_count("gender_6")
get_descriptive_count("gender_7")
get_descriptive_count("gender_8")
get_descriptive_count("gender_9")
get_descriptive_count("gender_10")
get_descriptive_count("gender_12") # No values for 11 ("I don't know")

# Sexuality
get_descriptive_count("sexuality")
get_descriptive_count("sexuality_2")
get_descriptive_count("sexuality_3")
get_descriptive_count("sexuality_4")
get_descriptive_count("sexuality_5")
get_descriptive_count("sexuality_6")
get_descriptive_count("sexuality_7")
get_descriptive_count("sexuality_8")
get_descriptive_count("sexuality_9")
get_descriptive_count("sexuality_10")

# Race/ethnicity
get_descriptive_count("race_ethnicity")

# Education
get_descriptive_count("education")

# Employment
get_descriptive_count("employment_employed")
get_descriptive_count("employment_unemployed")
get_descriptive_count("employment_student")
get_descriptive_count("employment_retired")
get_descriptive_count("employment_homemaker")
get_descriptive_count("employment_unable")
```

# RQ 1: Who self-identifies as autistic without a formal diagnosis?

## Descriptive Statistics

### Continuous Outcomes

```{r, results = "asis"}
# Pull names for continuous variables
continuous_outcomes <- df %>%
  select(age,
         starts_with("cati"),
         raads_sum,
         phq_dep, 
         phq_anx,
         whodas_sum, 
         wemwbs_sum,
         matches("support_.*_total"),
         starts_with("asis"),
         age_first_aware) %>%
  names()

# Print descriptives
map(
  .x = continuous_outcomes,
  .f = get_descriptive_continuous
)

# Function to analyze continuous variables
output_continuous <- function(x) {

  # Calculate test statistic
  test <- t_test(
    x = df,
    formula = as.formula(x %+% " ~ self_id"),
    order = c("TRUE", "FALSE") # Doesn't do anything important, but suppresses a warning
  )

  # Format output
  out <- tibble(
    statistic = "t(" %+% round(test$t_df) %+% ") = " %+% round(test$statistic, 2),
    p = test$p_value
  )

  return(out)
  
}

# Analyze continuous outcomes
rq1_continuous <- map(
  .x = continuous_outcomes,
  .f = output_continuous
) %>%
  set_names(continuous_outcomes) %>%
  bind_rows(.id = "x") 
```

### Nominal Outcomes

```{r, results = "asis"}
# Pull names for nominal variables
nominal_outcomes <- df %>%
  select(gender, 
         sexuality, 
         race_ethnicity_bin,
         starts_with("employment"), -employment_retired, # Expecteds too low for _retired
         raads_cutoff, 
         disability,
         support_current_mental_health:support_need_health,
         matches("^support_.*any$"),
         autism_dis, 
         starts_with("fam"),
         starts_with("friend")) %>%
  names()

# Print descriptives
map(
  .x = nominal_outcomes,
  .f = get_descriptive_count
)

# Function to analyze nominal variables
output_nominal <- function(x) {
  
  # Calculate test statistic
  test <- chisq_test(
    x = df,
    formula = as.formula(x %+% " ~ self_id")
  )

  # Format output
  out <- tibble(
    statistic = "X2(" %+% round(test$chisq_df) %+% ") = " %+% round(test$statistic, 2),
    p = test$p_value
  )

  return(out)
  
}

# Analyze nominal outcomes
rq1_nominal <- map(
  .x = nominal_outcomes,
  .f = output_nominal
) %>%
  set_names(nominal_outcomes) %>%
  bind_rows(.id = "x")
```

### Ordinal Outcomes

```{r, results = "asis"}
# Pull names for ordinal variables
ordinal_outcomes <- c("education_ord", "speaking_now", "speaking_child")

# Print descriptives
map(
  .x = ordinal_outcomes,
  .f = get_descriptive_count
)

# Function to analyze ordinal variables
output_ordinal <- function(x) {

  # Calculate test statistic
  test <- wilcox.test(
    as.numeric(df[[x]]) ~ df$self_id, 
    na.rm = T, 
    exact = F
  )
  
  # Format output
  out <- tibble(
    statistic = "U = " %+% round(test$statistic, 2),
    p = test$p.value
  )

  return(out)
  
}

# Analyze ordinal outcomes
rq1_ordinal <- map(
  .x = ordinal_outcomes,
  .f = output_ordinal
) %>%
  set_names(ordinal_outcomes) %>%
  bind_rows(.id = "x")
```

## Test Statistics and Adjusted _p_-Values

```{r, results = "asis"}
# Combine outputs
rq1 <- bind_rows(
  rq1_continuous,
  rq1_nominal,
  rq1_ordinal
)

# Adjust p-values
rq1_adjusted <- rq1 %>%
  mutate(p.adj = round(p.adjust(p, "BH"), 3)) %>%
  select(-p)

# Print
rq1_adjusted %>%
  kable() %>%
  kable_styling()
```

# Figures

## Autistic Traits

```{r}
df %>%
  select(response_id, self_id, starts_with("cati"), raads_sum) %>%
  pivot_longer(
    c(starts_with("cati"), raads_sum)
  ) %>%
  drop_na() %>%
  group_by(self_id, name) %>%
  summarize(mean = mean(value),
            se = sd(value) / sqrt(n()),
            .groups = "drop") %>%
  mutate(ci_high = mean + (1.96 * se),
         ci_low = mean - (1.96 * se),
         name = recode_factor(
           name, 
           "cati_com" = "Communication",
           "cati_rep" = "Repetitive Behaviors",
           "cati_rig" = "Cognitive Rigidity",
           "cati_sen" = "Sensory Sensitivity",
           "cati_soc" = "Social Interactions",
           "cati_cam" = "Social Camouflage",
           "cati_sum" = "CATI Sum",
           "raads_sum" = "RAADS-14 Sum"
         )) %>%
  ggplot(aes(self_id, mean, group = self_id)) +
  geom_col(aes(fill = self_id), position = "dodge") +
  geom_errorbar(aes(ymin = ci_low, ymax = ci_high), position = position_dodge(.9), width = .25) +
  scale_x_discrete(name = "Group", labels = c("Diagnosis", "Self-ID")) +
  scale_y_continuous(name = "Mean") +
  scale_fill_discrete(name = "Group", labels = c("Diagnosis", "Self-ID")) +
  ggtitle("Autistic Traits by Group") +
  theme_classic() +
  theme(legend.position = "none") +
  facet_wrap(~ name, ncol = 4, scales = "free")

ggsave(filename = here("Figures", "Autistic Traits by Group.png"),
       width = 8,
       height = 5,
       units = "in",
       dpi = 300)
```

## Supports

```{r}
df %>%
  select(response_id, self_id, starts_with("support"), -ends_with("total"), -ends_with("any")) %>%
  pivot_longer(
    starts_with("support")
  ) %>%
  group_by(self_id, name) %>%
  summarize(prop = mean(value),
            se = sqrt((prop * (1 - prop)) / n()),
            .groups = "drop") %>%
  mutate(ci_high = prop + (1.96 * se),
         ci_low = prop - (1.96 * se),
         vartype = case_when(
           grepl("current", name) ~ "Current",
           grepl("past", name) ~ "Past",
           grepl("need", name) ~ "Unmet Need"
         ) %>%
           factor(levels = c("Past", "Current", "Unmet Need")),
         support = name %>%
           gsub("^([^_]*_){2}", "", .) %>%
           recode_factor(
             "case_management" = "Case Management",
             "functional" = "Functional\n(e.g., Speech\nTherapy)",
             "health" = "Health\n(e.g., Specialized\nMedical Care)",
             "mental_health" = "Mental Health\n(e.g., Therapy)",
             "social" = "Social\n(e.g., Social\nSkills Training)",
             "vocational" = "Vocational\n(e.g., Job Training)"
           )) %>%
  ggplot(aes(support, prop, group = self_id)) +
  geom_col(aes(fill = self_id), position = "dodge") +
  geom_errorbar(aes(ymin = ci_low, ymax = ci_high), position = position_dodge(.9), width = .25) +
  scale_x_discrete(name = "Support Type") +
  scale_y_continuous(name = "Percent", labels = scales::percent) +
  scale_fill_discrete(name = "Group", labels = c("Diagnosis", "Self-ID")) +
  ggtitle("Support Access and Need by Group") +
  theme_classic() +
  facet_wrap(~ vartype, ncol = 1, scales = "free")

ggsave(filename = here("Figures", "Supports by Group.png"),
       width = 8,
       height = 5,
       units = "in",
       dpi = 300)
```

## Psychosocial Outcomes

```{r}
# Plotting PHQ-2
dep_id <- df %>%
  filter(self_id == T) %>%
  drop_na(phq_dep) %>%
  pull(phq_dep)

dep_dx <- df %>%
  filter(self_id == F) %>%
  drop_na(phq_dep) %>%
  pull(phq_dep)

dep_norm <- tibble(
  phq_dep = 0:6,
  pr = c(24.3, 59.9, 80.9, 93.0, 96.8, 98.7, 99.6),
  radius = NA
)

for(i in 1:nrow(dep_norm)) {
  
  if(i == 1) {
    
    dep_norm$radius[i] <- dep_norm$pr[i] * 2
    
  } else {
    
    dep_norm$radius[i] <- (dep_norm$pr[i] - dep_norm$radius[i - 1]) + dep_norm$pr[i]
    
  }
  
}

dep_norm <- dep_norm %>%
  mutate(
    radius_lag = lag(radius, 1, default = 0),
    pr_size = radius - radius_lag,
    n = pr_size * 10
  )

dep_norm_temp <- rep(dep_norm$phq_dep, dep_norm$n)

dep_norm_to_plot <- bind_rows(
  tibble(group = "Autistic: Diagnosed", phq_dep = dep_dx), 
  tibble(group = "Autistic: Self-Identify", phq_dep = dep_id),
  tibble(group = "General Population", phq_dep = dep_norm_temp)
) 

dep_plot <- dep_norm_to_plot %>%
  ggplot() +
  geom_density(aes(phq_dep, group = group, fill = group), alpha = .25, bw = 1) +
  scale_x_continuous(name = "PHQ-2 Sum Score") +
  scale_y_continuous(name = "Density", labels = scales::percent) +
  scale_fill_discrete(name = "Group") +
  ggtitle("Depression Symptom Severity by Group") +
  theme_classic()
```

```{r}
# Plotting GAD-2
anx_id <- df %>%
  filter(self_id == T) %>%
  drop_na(phq_anx) %>%
  pull(phq_anx)

anx_dx <- df %>%
  filter(self_id == F) %>%
  drop_na(phq_anx) %>%
  pull(phq_anx)

anx_norm <- tibble(
  phq_anx = 0:6,
  pr = c(28.2, 66.4, 83.9, 93.5, 96.8, 98.5, 99.5),
  radius = NA
)

for(i in 1:nrow(anx_norm)) {
  
  if(i == 1) {
    
    anx_norm$radius[i] <- anx_norm$pr[i] * 2
    
  } else {
    
    anx_norm$radius[i] <- (anx_norm$pr[i] - anx_norm$radius[i - 1]) + anx_norm$pr[i]
    
  }
  
}

anx_norm <- anx_norm %>%
  mutate(
    radius_lag = lag(radius, 1, default = 0),
    pr_size = radius - radius_lag,
    n = pr_size * 10
  )

anx_norm_temp <- rep(anx_norm$phq_anx, anx_norm$n)

anx_norm_to_plot <- bind_rows(
  tibble(group = "Autistic: Diagnosed", phq_anx = anx_dx), 
  tibble(group = "Autistic: Self-Identify", phq_anx = anx_id),
  tibble(group = "General Population", phq_anx = anx_norm_temp)
)

anx_plot <- anx_norm_to_plot %>%
  ggplot() +
  geom_density(aes(phq_anx, group = group, fill = group), alpha = .25, bw = 1) +
  scale_x_continuous(name = "GAD-2 Sum Score") +
  scale_y_continuous(name = "Density", labels = scales::percent) +
  scale_fill_discrete(name = "Group") +
  ggtitle("Anxiety Symptom Severity by Group") +
  theme_classic()
```

```{r}
# Plotting WEMWBS
wemwbs_id <- df %>%
  filter(self_id == T) %>%
  drop_na(wemwbs_sum) %>%
  pull(wemwbs_sum)

wemwbs_dx <- df %>%
  filter(self_id == F) %>%
  drop_na(wemwbs_sum) %>%
  pull(wemwbs_sum)

wemwbs_norm <- rnorm(1000, 48.9, 9.8)

wemwbs_norm_to_plot <- bind_rows(
  tibble(group = "Autistic: Diagnosed", wemwbs = wemwbs_dx), 
  tibble(group = "Autistic: Self-Identify", wemwbs = wemwbs_id),
  tibble(group = "General Population", wemwbs = wemwbs_norm)
) 

wemwbs_plot <- wemwbs_norm_to_plot %>%
  ggplot() +
  geom_density(aes(wemwbs, group = group, fill = group), alpha = .25, bw = 5) +
  scale_x_continuous(name = "WEMWBS Sum Score") +
  scale_y_continuous(name = "Density", labels = scales::percent) +
  scale_fill_discrete(name = "Group") +
  ggtitle("Mental Well-Being by Group") +
  theme_classic()
```

```{r}
# Combined plot
combined_plot <- 
  dep_plot +
  anx_plot + 
  wemwbs_plot +
  plot_layout(
    ncol = 1,
    guides = "collect"
  )

combined_plot

ggsave(plot = combined_plot, 
       filename = here("Figures", "Psychological Outcome Distributions Group.png"),
       width = 5.5,
       height = 7,
       units = "in",
       dpi = 300)
```

# Additional Analyses

## How many self-identifiers want a diagnosis?

With Wald confidence intervals (normal approximation) and Wilson for reference

```{r}
df %>%
  filter(self_id) %>%
  count(nodx_want) %>%
  mutate(
    prop = n / sum(n),
    total_n = sum(n),
    se = sqrt((prop * (1 - prop)) / total_n),
    ci_high = prop + (1.96 * se),
    ci_low = prop - (1.96 * se)
  )

DescTools::MultinomCI(c(101, 10, 36), method = "wald")

DescTools::MultinomCI(c(101, 10, 36), method = "wilson")
```

# Content Analysis: Code Frequencies

```{r}
# Q1
df %>%
  pivot_longer(
    cols = starts_with("q1")
  ) %>%
  group_by(name) %>%
  summarize(
    n = sum(value, na.rm = T),
    pct = percent(mean(value, na.rm = T), .1)
  ) %>%
  kable() %>%
  kable_styling()

# Q2
df %>%
  pivot_longer(
    cols = starts_with("q2")
  ) %>%
  group_by(name) %>%
  summarize(
    n = sum(value, na.rm = T),
    pct = percent(mean(value, na.rm = T), .1)
  ) %>%
  kable() %>%
  kable_styling()

# Q3
df %>%
  mutate(
    q3_healthcare_parent = q3_healthcare_professionals | q3_healthcare_other
  ) %>%
  pivot_longer(
    cols = starts_with("q3")
  ) %>%
  group_by(name) %>%
  summarize(
    n = sum(value, na.rm = T),
    pct = percent(mean(value, na.rm = T), .1)
  ) %>%
  kable() %>%
  kable_styling()

# Q4
df %>%
  pivot_longer(
    cols = starts_with("q4")
  ) %>%
  group_by(name) %>%
  summarize(
    n = sum(value, na.rm = T),
    pct = percent(mean(value, na.rm = T), .1)
  ) %>%
  kable() %>%
  kable_styling()

# By group
q4_cols <- df %>%
  select(starts_with("q4")) %>%
  names()

q4_comparison <- tibble()

for(x in q4_cols) {
  
  table <- table(df$self_id, df[[x]])
  
  sum_self_id <- sum(df[[x]][df$self_id], na.rm = T)
  mean_self_id <- percent(mean(df[[x]][df$self_id], na.rm = T), .1)
  
  sum_dx <- sum(df[[x]][!df$self_id], na.rm = T)
  mean_dx <- percent(mean(df[[x]][!df$self_id], na.rm = T), .1)

  chisq <- chisq.test(table)
  
  expected <- chisq$expected
  
  if(min(expected) >= 5) {
    
    statistic <- chisq$statistic
    p_value <- chisq$p.value
    
  } else {
    
    statistic <- NA
    p_value <- NA
    
  }
  
  new_row <- tibble(
    var = x,
    self_id = sum_self_id %+% " (" %+% mean_self_id %+% ")",
    dx = sum_dx %+% " (" %+% mean_dx %+% ")",
    statistic = statistic,
    p_value = p_value
  )
  
  q4_comparison <- bind_rows(q4_comparison, new_row)
  
}

q4_comparison %>%
  mutate(p_adj = p.adjust(p_value, "BH")) %>%
  kable() %>%
  kable_styling()

# Q5
df %>%
  pivot_longer(
    cols = starts_with("q5")
  ) %>%
  group_by(name) %>%
  summarize(
    n = sum(value, na.rm = T),
    pct = percent(mean(value, na.rm = T), .1)
  ) %>%
  kable() %>%
  kable_styling()

# By group
q5_cols <- df %>%
  select(starts_with("q5")) %>%
  names()

q5_comparison <- tibble()

for(x in q5_cols) {
  
  table <- table(df$self_id, df[[x]])
  
  sum_self_id <- sum(df[[x]][df$self_id], na.rm = T)
  mean_self_id <- percent(mean(df[[x]][df$self_id], na.rm = T), .1)
  
  sum_dx <- sum(df[[x]][!df$self_id], na.rm = T)
  mean_dx <- percent(mean(df[[x]][!df$self_id], na.rm = T), .1)

  chisq <- chisq.test(table)
  
  expected <- chisq$expected
  
  if(min(expected) >= 5) {
    
    statistic <- chisq$statistic
    p_value <- chisq$p.value
    
  } else {
    
    statistic <- NA
    p_value <- NA
    
  }
  
  new_row <- tibble(
    var = x,
    self_id = sum_self_id %+% " (" %+% mean_self_id %+% ")",
    dx = sum_dx %+% " (" %+% mean_dx %+% ")",
    statistic = statistic,
    p_value = p_value
  )
  
  q5_comparison <- bind_rows(q5_comparison, new_row)
  
}

q5_comparison %>%
  mutate(p_adj = p.adjust(p_value, "BH")) %>%
  kable() %>%
  kable_styling()
```
